refactor: Update background image imports in Login and Register pages
